requirements.txt
cryptography==39.0.1
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.15.0
six==1.15.0
